+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.34.0](https://github.com/anzusystems/common-admin/compare/1.33.0...1.34.0) (unreleased)
+
+### Features
+* **chore:** updated dependencies
+
+### BREAKING CHANGES
+* **dependencies:** new vue-i18n required version `vue-i18n@9.6.2`
+* **removed:** `ACheckboxSimple`, use `VCheckboxBtn`
+* **changed:** `useSubjectSelect` function arguments
+
 ## [1.33.0](https://github.com/anzusystems/common-admin/compare/1.32.0...1.33.0) (2023-10-27)
 
 ### Features